Move genreListItem to a separate file
