+++ lib/movie_card/genres_list.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+part 'genres_list_item.dart';
+
 class GenresList extends StatelessWidget {
   static const String title = 'Genres';
   static const int titleFlex = 1;
@@ -49,41 +51,6 @@ class GenresList extends StatelessWidget {
             ),
           ),
         ],
-      ),
-    );
-  }
-}
-
-class _GenreListItem extends StatelessWidget {
-  const _GenreListItem({
-    required this.genre,
-  });
-
-  static const double borderRadius = 18;
-  static const double horizontalButtonPadding = 3;
-
-  final String genre;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(
-        horizontal: horizontalButtonPadding,
-      ),
-      child: OutlinedButton(
-        onPressed: () {},
-        style: OutlinedButton.styleFrom(
-          side: BorderSide(
-            color: Theme.of(context).colorScheme.secondary,
-          ),
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(borderRadius),
-          ),
-        ),
-        child: Text(
-          genre,
-          style: Theme.of(context).textTheme.bodySmall,
-        ),
       ),
     );
   }
